examples/quickstart/modal/requirements.txt
moondream
pillow
modal